+++ ft_memccpy.c
@@ -6,17 +6,17 @@
 /*   By: jpceia <jpceia@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/08 21:47:05 by jpceia            #+#    #+#             */
-/*   Updated: 2021/02/09 00:11:59 by jpceia           ###   ########.fr       */
+/*   Updated: 2021/02/09 21:38:42 by jpceia           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	*ft_memccpy(void *dest, const void *src, int c, int n)
+void	*ft_memccpy(void *dest, const void *src, int c, size_t n)
 {
 	char	*ptr_dest;
 	char	*ptr_src;
-	int				index;
+	size_t	index;
 
 	ptr_dest = (char *)dest;
 	ptr_src = (char *)src;
