+++ ft_memmove.c
@@ -6,18 +6,18 @@
 /*   By: jpceia <jpceia@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/09 15:09:24 by jpceia            #+#    #+#             */
-/*   Updated: 2021/02/09 15:18:14 by jpceia           ###   ########.fr       */
+/*   Updated: 2021/02/09 21:38:56 by jpceia           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void *ft_memmove(void *dest, const void *src, int n)
+void *ft_memmove(void *dest, const void *src, size_t n)
 {
 	char	*ptr_dest;
 	char	*ptr_src;
 	char	*holder;
-	int		index;
+	size_t	index;
 
 	ptr_dest = (char *)dest;
 	ptr_src = (char *)src;
